Modularized web_exist.py with updated README
